+++ src/bandcamp/pages/page-music.js
@@ -1,6 +1,6 @@
-import { contentChangeWithPolling, createDatalistFromArray, createElementFromHTML, input, selectElementWithContent, setDataAttribute } from "../modules/html.js";
-import { containsOneOf, splitString, isEmptyArray, countOccurrences, removeBrackets } from "../modules/utils.js";
-import { getBandPhotoSrc, getReleasesData } from "./html.js";
+import { contentChangeWithPolling, createDatalistFromArray, createElementFromHTML, input, selectElementWithContent, setDataAttribute } from "../../modules/html.js";
+import { containsOneOf, splitString, isEmptyArray, countOccurrences, removeBrackets } from "../../modules/utils.js";
+import { getBandPhotoSrc, getReleasesData } from "../modules/html.js";
 
 // Setup logic for BC music page
 export function setupPageMusic() {